Following guidelines of Hugo version 0.146.0
From 0.146.0, Hugo authors has implemented a breaking change on how layout template are arranged. TL;DR, you can find all the changes listed here : https://gohugo.io/templates/new-templatesystem-overview/.

If you are still reading, this basically re-arranges the contents of layouts folder into a flat structure; with all the templates residing in layouts or one level deeper at most (unless there is a need for per-page specific layout). This also changed template lookup order across all template types.

Additionally, you can find the folder structure which is expected with Hugo >=0.146.0. You may need to change your folder structure if you have extended this theme. Nevertheless, Hugo will try to be backwards compatible for the time-being, but it is a matter of time when this compatibility is broken.

If you are not using Hugo version >=0.146.0 for building this theme, I would urge you to do so. Debian packages are fashionably out-of-date, and is still struck with 0.140.0 (as of September 2025). You might want to at-least try to get 0.146.0 binary for building.

If you have any issues, you know where to post your grievance ;-)