Move rss.xml in the layouts/ directory
Initially it was in layouts/posts/ where it was ignored, so I moved it
to layouts/_default which is the previous location where it was supposed
to be. It changed in version 0.146. As explained in [1], everything
in the _default/ directory must be moved to the root of the layouts/
directory.

[1]: https://gohugo.io/templates/new-templatesystem-overview/